ruo-rag-langserver/packages/patent-rag/pyproject.toml
[tool.poetry]
name = "patent-rag"
version = "0.1.0"
description = ""
authors = ["Ruo Lee <comsa33@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.16"
langchain-community = "^0.0.34"
langchainhub = "^0.1.15"
openai = "^1.24.0"
langchain-openai = "^0.1.4"
python-dotenv = "^1.0.1"
langchain-elasticsearch = "^0.1.3"
huggingface-hub = "^0.22.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
langchain-cli = ">=0.0.21"

[tool.langserve]
export_module = "patent_rag"
export_attr = "chain"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
